+++ TradeGates/Utils/BinanceHelpers.py
@@ -1,104 +1,9 @@
+import json
 from datetime import datetime
 
 import pandas as pd
 
-from TradeGates.Exchanges.BinanceExchange import BinanceExchange
-from TradeGates.Utils import DataHelpers
-
-
-def isOrderDataValid(order: DataHelpers.OrderData):
-    if order.orderType not in BinanceExchange.spotOrderTypes:
-        return False
-
-    if order.side not in ['BUY', 'SELL']:
-        return False
-
-    if order.newOrderRespType not in [None, 'ACK', 'RESULT', 'FULL']:
-        return False
-
-    if order.timeInForce not in [None, 'GTC', 'IOC', 'FOK']:
-        return False
-
-    if order.orderType == 'LIMIT':
-        if not (order.timeInForce is None or order.quantity is None or order.price is None):
-            return True
-
-    elif order.orderType == 'MARKET':
-        if not (order.quantity is None and order.quoteOrderQty is None):
-            return True
-
-    elif order.orderType in ['STOP_LOSS', 'TAKE_PROFIT']:
-        if not (order.quantity is None or order.stopPrice is None):
-            return True
-
-    elif order.orderType in ['STOP_LOSS_LIMIT', 'TAKE_PROFIT_LIMIT']:
-        if not (
-                order.timeInForce is None or order.quantity is None or order.price is None or order.stopPrice is None):
-            return True
-
-    elif order.orderType == 'LIMIT_MAKER':
-        if not (order.quantity is None or order.price is None):
-            return True
-
-    return False
-
-
-def isFuturesOrderDataValid(order: DataHelpers.futuresOrderData):
-    if order.side not in ['BUY', 'SELL']:
-        return False
-
-    if order.orderType not in BinanceExchange.futuresOrderTypes:
-        return False
-
-    if order.positionSide not in [None, 'BOTH', 'LONG', 'SHORT']:
-        return False
-
-    if order.timeInForce not in [None, 'GTC', 'IOC', 'FOK', 'GTX']:
-        return False
-
-    if order.workingType not in [None, 'MARK_PRICE', 'CONTRACT_PRICE']:
-        return False
-
-    if order.newOrderRespType not in [None, 'ACK', 'RESULT']:
-        return False
-
-    if order.closePosition not in [None, True, False]:
-        return False
-
-    if order.callbackRate is not None and not (0.1 <= order.callbackRate <= 5):
-        return False
-
-    if order.priceProtect not in [None, True, False]:
-        return False
-
-    if order.closePosition is True and order.quantity is not None:
-        return False
-
-    if order.reduceOnly not in [None, True, False]:
-        return False
-
-    if order.closePosition is True and order.reduceOnly is True:
-        return False
-
-    if order.orderType == 'LIMIT':
-        if not (order.timeInForce is None or order.quantity is None or order.price is None):
-            return True
-
-    elif order.orderType == 'MARKET':
-        if order.quantity is not None:
-            return True
-
-    elif order.orderType in ['STOP', 'TAKE_PROFIT']:
-        if not (order.quantity is None or order.price is None or order.stopPrice is None):
-            return True
-
-    elif order.orderType in ['STOP_MARKET', 'TAKE_PROFIT_MARKET']:
-        if order.stopPrice is not None:
-            return True
-
-    elif order.orderType == 'TRAILING_STOP_MARKET':
-        if order.callbackRate is not None:
-            return True
+from Utils import DataHelpers
 
 
 def getSpotOrderAsDict(order: DataHelpers.OrderData):
@@ -189,10 +94,10 @@ def getFuturesOrderAsDict(order: DataHelpers.futuresOrderData, allStr=False):
     return params
 
 
-def getKlinesDesiredOnlyCols(data):
+def getKlinesDesiredOnlyCols(data, desiredIndexes):
     finalDataArray = []
     for datum in data:
-        finalDataArray.append([datum[index] for index in BinanceExchange.desiredCandleDataIndexes])
+        finalDataArray.append([datum[index] for index in desiredIndexes])
     return finalDataArray
 
 
@@ -203,9 +108,9 @@ def klinesConvertToPandas(outArray):
     return df
 
 
-def klinesConvertDate(data):
+def klinesConvertDate(data, timeColIdxs):
     for datum in data:
-        for idx in BinanceExchange.timeIndexesInCandleData:
+        for idx in timeColIdxs:
             datum[idx] = datetime.fromtimestamp(float(datum[idx]) / 1000)
 
 
@@ -214,7 +119,7 @@ def extractSymbolInfoFromFilters(symbolFilters, tickerPrice):
     for symbolFilter in symbolFilters:
         if symbolFilter['filterType'] == 'LOT_SIZE':
             params['minQuantity'] = float(symbolFilter['minQty'])
-            params['stepQuantity'] = float(symbolFilter['stepSize'])
+            params['precisionStep'] = float(symbolFilter['stepSize'])
             params['minQuoteQuantity'] = tickerPrice * params['minQuantity']
 
         if symbolFilter['filterType'] == 'PRICE_FILTER':
@@ -222,10 +127,10 @@ def extractSymbolInfoFromFilters(symbolFilters, tickerPrice):
     return params
 
 
-def makeBatchOrderData(self, futuresOrderDatas):
+def makeBatchOrderData(futuresOrderDatas):
     batchOrders = []
     for order in futuresOrderDatas:
-        orderAsDict = BinanceHelpers.getFuturesOrderAsDict(order, allStr=True)
+        orderAsDict = getFuturesOrderAsDict(order, allStr=True)
         orderAsDict['type'] = orderAsDict.pop('ordertype')
 
         orderJSON = json.dumps(orderAsDict)
